Refactor project and split code into separate files/folders using ES Modules.
