+++ index.html
@@ -8,7 +8,7 @@
   <link rel="apple-touch-icon" href="./favicon.ico">
   <title>ProjectManager</title>
   <link rel="stylesheet" href="app.css" />
-  <script src="dist/app.js" defer></script>
+  <script src="dist/app.js" type="module"></script>
 </head>
 <body>
 <template id="project-input">
