+++ src/decorators/autobind.ts
@@ -1,0 +1,14 @@
+// Autobind Decorator
+//eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function Autobind(_: any, _2: string, descriptor: PropertyDescriptor): PropertyDescriptor {
+  const originalMethod = descriptor.value;
+  const adjustedDescriptor: PropertyDescriptor = {
+    configurable: true,
+    enumerable: false,
+    get() {
+      const boundFn = originalMethod.bind(this);
+      return boundFn;
+    },
+  };
+  return adjustedDescriptor;
+}
