+++ src/state/project-state.ts
@@ -1,0 +1,60 @@
+import { Project, ProjectStatus } from "../models/project.js";
+
+type Listener<T> = (items: Array<T>) => void;
+
+// State Management class
+class State<T> {
+  protected listeners: Array<Listener<T>> = [];
+
+  addListener(listenerFn: Listener<T>) {
+    this.listeners.push(listenerFn);
+  }
+}
+// ProjectState Management class
+export class ProjectState extends State<Project> {
+  private projects: Array<Project> = [];
+  private static instance: ProjectState;
+
+  private constructor() {
+    super();
+  }
+
+  static getInstance() {
+    if (this.instance) {
+      return this.instance;
+    }
+
+    this.instance = new ProjectState();
+    return this.instance;
+  }
+
+  addProject(title: string, description: string, numOfPeople: number) {
+    const newProject = new Project(
+      Math.random().toString(),
+      title,
+      description,
+      numOfPeople,
+      ProjectStatus.ACTIVE
+    );
+
+    this.projects.push(newProject);
+
+    this.updateListeners();
+  }
+
+  moveProject(projectId: string, newStatus: ProjectStatus) {
+    const project = this.projects.find((p) => p.id === projectId);
+    if (project && project.status !== newStatus) {
+      project.status = newStatus;
+      this.updateListeners();
+    }
+  }
+
+  private updateListeners() {
+    for (const listenerFn of this.listeners) {
+      listenerFn(this.projects.slice());
+    }
+  }
+}
+
+export const projectState = ProjectState.getInstance();
